優化性能並修正高 DPI 下的圖形問題

+++ ConcreteUI/Controls/GroupBox.cs
@@ -65,7 +65,7 @@ namespace ConcreteUI.Controls
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private LayoutVariable GetContentLayoutReferenceCore(nuint property)
         {
-            ref LayoutVariable? variable = ref UnsafeHelper.AddTypedOffset(ref UnsafeHelper.GetArrayDataReference(_contentLayoutReferences), (nuint)property);
+            ref LayoutVariable? variable = ref UnsafeHelper.AddTypedOffset(ref UnsafeHelper.GetArrayDataReference(_contentLayoutReferences), property);
             if (variable is null)
             {
                 WeakReference<GroupBox>? reference = InterlockedHelper.Read(ref _reference);
